move parallel tests to full .NET
apparently we're using multicast delegates which are not supported on .NET Standard. see dotnet/corefx#5940
