+++ Testing/unittest/Writing/StoreWriterTests.cs
@@ -80,72 +80,6 @@ namespace VDS.RDF.Writing
         }
 
         [Fact]
-        public void WritingNQuadsSingleThreaded()
-        {
-            this.WritingNQuadsActual();
-        }
-
-        private void WritingNQuadsActual()
-        {
-            this.TestWriter(new NQuadsWriter(NQuadsSyntax.Original), new NQuadsParser(NQuadsSyntax.Original), true);
-        }
-
-        [Fact]
-        public void WritingNQuadsMixedSingleThreaded()
-        {
-            this.WritingNQuadsMixedActual();
-        }
-
-        private void WritingNQuadsMixedActual()
-        {
-            this.TestWriter(new NQuadsWriter(NQuadsSyntax.Original), new NQuadsParser(NQuadsSyntax.Rdf11), true);
-        }
-
-        [Fact]
-        public void WritingNQuadsMixedBadSingleThreaded()
-        {
-            Assert.Throws<RdfParseException>(() => this.WritingNQuadsMixedBadActual());
-        }
-
-        private void WritingNQuadsMixedBadActual()
-        {
-            this.TestWriter(new NQuadsWriter(NQuadsSyntax.Rdf11), new NQuadsParser(NQuadsSyntax.Original), true);
-        }
-
-        [Fact]
-        public void WritingNQuads11SingleThreaded()
-        {
-            this.WritingNQuads11Actual();
-        }
-
-        private void WritingNQuads11Actual()
-        {
-            this.TestWriter(new NQuadsWriter(NQuadsSyntax.Rdf11), new NQuadsParser(NQuadsSyntax.Rdf11), true);
-        }
-
-        [Fact]
-        public void WritingTriGSingleThreaded()
-        {
-            this.WritingTriGActual();
-        }
-
-        private void WritingTriGActual()
-        {
-            this.TestWriter(new TriGWriter(), new TriGParser(), true);
-        }
-
-        [Fact]
-        public void WritingTriGUncompressedSingleThreaded()
-        {
-            this.WritingTriGUncompressedActual();
-        }
-
-        private void WritingTriGUncompressedActual()
-        {
-            this.TestWriter(new TriGWriter(), new TriGParser(), true, WriterCompressionLevel.None);
-        }
-
-        [Fact]
         public void WritingTriX()
         {
             this.TestWriter(new TriXWriter(), new TriXParser(), false);
